ADD :: JoinContents 모듈화
