+++ src/components/auth/Join/JoinContents/style.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const InputWrap = styled.div`
+  display: flex;
+  flex-direction: column;
+  margin: 4rem 0rem;
+  gap: 36px;
+  width: 64%;
+`;
