+++ src/type/common/input.type.ts
@@ -3,5 +3,6 @@ export interface InputPropsType {
   type: string;
   name: string;
   placeholder: string;
-  onChange: React.ChangeEventHandler<HTMLInputElement>;
+  onChange?: React.ChangeEventHandler<HTMLInputElement>;
+  onFocus?: React.FocusEventHandler<HTMLInputElement>;
 }
